pc/by - refactor to create strategies directory
